+++ test/support/agent_case.ex
@@ -153,11 +153,11 @@ defmodule Omni.Agent.AgentCase do
 
       defp collect_events_loop(agent_pid, acc, timeout) do
         receive do
-          {:agent, ^agent_pid, :continue, data} ->
-            collect_events_loop(agent_pid, [{:continue, data} | acc], timeout)
+          {:agent, ^agent_pid, :turn, {:stop, _} = data} ->
+            Enum.reverse([{:turn, data} | acc])
 
-          {:agent, ^agent_pid, :stop, data} ->
-            Enum.reverse([{:stop, data} | acc])
+          {:agent, ^agent_pid, :turn, {:continue, _} = data} ->
+            collect_events_loop(agent_pid, [{:turn, data} | acc], timeout)
 
           {:agent, ^agent_pid, :error, response} ->
             Enum.reverse([{:error, response} | acc])